Fix error if the first container is closed
If the first container in the group is closed, it causes a null pointer
error. The `ui` is null for a closed container because it can't get the
`ui` for a closed container.

Previously, it gets the ui only from the first container. To fix this
problem, now it tries to get the ui from all containers.
If no container opens, it does nothing.